- FIX: Just some isset checks on my $_POST data. Thanks Tom.

+++ admin.actions.php
@@ -22,15 +22,15 @@ function hmbkp_option_save() {
 	global $hmbkp_errors;
 	$hmbkp_errors = new WP_Error;
 	
-	if( ! (bool) $_POST['hmbkp_automatic'] )
+	if( isset( $_POST['hmbkp_automatic'] ) && ! (bool) $_POST['hmbkp_automatic'] )
 		update_option('hmbkp_disable_automatic_backup', 'true' );
 	else 
 		delete_option('hmbkp_disable_automatic_backup');
 	
-	if( $_POST['hmbkp_what_to_backup'] == 'files only' ) {
+	if( isset( $_POST['hmbkp_what_to_backup'] ) && $_POST['hmbkp_what_to_backup'] == 'files only' ) {
 		update_option('hmbkp_files_only', 'true' );
 		delete_option('hmbkp_database_only');
-	} elseif( $_POST['hmbkp_what_to_backup'] == 'database only' ) {
+	} elseif( isset( $_POST['hmbkp_what_to_backup'] ) && $_POST['hmbkp_what_to_backup'] == 'database only' ) {
 		update_option('hmbkp_database_only', 'true' );
 		delete_option('hmbkp_files_only');
 	} else {
@@ -38,22 +38,22 @@ function hmbkp_option_save() {
 		delete_option('hmbkp_files_only');
 	}
 	
-	if( $max_backups = intval( $_POST['hmbkp_backup_number'] ) ) {
+	if( isset( $_POST['hmbkp_max_backups'] ) && $max_backups = intval( $_POST['hmbkp_backup_number'] ) ) {
 		update_option('hmbkp_max_backups', intval( $_POST['hmbkp_backup_number'] ) );
 	} else {
-		delete_option('hmbkp_max_backups', intval( $_POST['hmbkp_backup_number'] ) );
+		delete_option( 'hmbkp_max_backups' );
 		$hmbkp_errors->add( 'invalid_no_backups', __("You have entered an invalid number of backups.") );
 	}
 	
-	if( !is_email( $_POST['hmbkp_email_address'] ) && !empty( $_POST['hmbkp_email_address'] ) ) {
+	if( isset( $_POST['hmbkp_email_address'] ) && !is_email( $_POST['hmbkp_email_address'] ) && !empty( $_POST['hmbkp_email_address'] ) ) {
 			$hmbkp_errors->add( 'invalid_email', __("You have entered an invalid email address.") );
-	} elseif( !empty( $_POST['hmbkp_email_address'] ) ) {
+	} elseif( isset( $_POST['hmbkp_email_address'] ) && !empty( $_POST['hmbkp_email_address'] ) ) {
 		update_option( 'hmbkp_email_address', $_POST['hmbkp_email_address'] );
 	} else {
 		delete_option( 'hmbkp_email_address' );
 	}
 	
-	if( !empty( $_POST['hmbkp_excludes'] ) ) {
+	if( isset( $_POST['hmbkp_excludes'] ) && !empty( $_POST['hmbkp_excludes'] ) ) {
 		update_option('hmbkp_excludes', $_POST['hmbkp_excludes'] );
 		delete_transient('hmbkp_estimated_filesize');
 	} else {
